objects: Fix related model that was pointing at the wrong model
When looking at the detail of an object (e.g. /browse/app/model/objects/42),
the related models were incorrectly displayed.  It showed all related models
as being the model currently displayed, so it showed this:

   Appears in "XX" in the following <model>

instead of:

   Appears in "XX" in the following <related model>

and there was the same issue for the URLs leading to related objects.